third_party/blink/web_tests/wpt_internal/js-self-profiling/foreign-script-cors-allow.sub.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>

  <script src="resources/profile-utils.js"></script>

  <script id="foreign-script" crossorigin src="http://{{hosts[][www1]}}:{{ports[http][0]}}/wpt_internal/js-self-profiling/resources/external-script.js"></script>
</head>
<body>
  <script>
    promise_test(async t => {
      assert_not_equals(window.externalScriptFunction, undefined);
      const trace = await ProfileUtils.profileFunction(externalScriptFunction);

      const foreignScriptUrl = document.querySelector('#foreign-script').src;
      assert_true(ProfileUtils.containsFrame(trace, {
        name: 'externalScriptFunction',
        resourceId: trace.resources.indexOf(foreignScriptUrl),
      }), 'foreign frame is present');

    }, 'function from foreign-origin script with CORS is included in trace');
  </script>
</body>
</html>
